fix x86_64 compile issue

+++ bank_test.c
@@ -285,7 +285,7 @@ double find_read_time(void *_a, void *_b, double threshold)
     uint64_t start_ticks, end_ticks, ticks;
     uint64_t min_ticks, max_ticks, sum_ticks;
     double avg_ticks;
-    uint64_t sum = 0;
+    int sum = 0;
 	int ticks_array[MAX_OUTER_LOOP];
 	
     assert((uintptr_t)(a) == (uintptr_t)(_a));
@@ -326,7 +326,7 @@ double find_read_time(void *_a, void *_b, double threshold)
         assert(*(uint64_t *)_b == 0);
         // TODO: Why is sum not zero?
         if (sum != 0)
-            printf("i=%d, *a=%ld, *b=%ld, sum=%ld\n",
+            printf("i=%d, *a=%ld, *b=%ld, sum=%d\n",
 				   i, *(uint64_t *)(_a), *(uint64_t *)(_b), sum);
         assert(sum == 0);
